moved JSRSimulation code into simulation
